requirements.txt
boto3==1.26.151
brotlipy==0.7.0
gitpython==3.1.31
huggingface-hub==0.15.1
imutils==0.5.4
mkl-fft==1.3.6
mkl-random==1.2.2
pip-chill==1.0.3
pyopenssl==23.0.0
pysocks==1.7.1
roboflow==1.0.9
sahi==0.11.14
tensorboard==2.13.0
thop==0.1.1.post2209072238
torchaudio==2.0.2
ultralytics==8.0.115
vietocr==0.3.2
win-inet-pton==1.1.0
boto3==1.26.151
brotlipy==0.7.0
gitpython==3.1.31
huggingface-hub==0.15.1
imutils==0.5.4
mkl-fft==1.3.6
mkl-random==1.2.2
pip-chill==1.0.3
pyopenssl==23.0.0
pysocks==1.7.1
roboflow==1.0.9
sahi==0.11.14
tensorboard==2.13.0
thop==0.1.1.post2209072238
torchaudio==2.0.2
ultralytics==8.0.115
vietocr==0.3.2
win-inet-pton==1.1.0
